pyproject.toml
[project]
name = "imagegen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "google>=3.0.0",
    "google-genai>=1.46.0",
    "google-generativeai>=0.8.5",
    "mutagen>=1.47.0",
    "numpy>=2.3.4",
    "pillow>=12.0.0",
]
